server/requirements.txt
-i https://pypi.org/simple
alembic==1.13.1; python_version >= '3.8'
aniso8601==9.0.1
asttokens==2.4.1
attrs==23.2.0; python_version >= '3.7'
bcrypt==4.1.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.7'
exceptiongroup==1.2.0; python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
faker==23.2.1; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-marshmallow==0.14.0
flask-migrate==3.1.0; python_version >= '3.6'
flask-restful==0.3.9
flask-restplus==0.13.0
flask-serializer==0.0.5.1; python_version >= '2.7'
flask-sqlalchemy==3.0.3; python_version >= '3.7'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0; python_version >= '3.5'
importlib-metadata==6.0.0; python_version >= '3.7'
importlib-resources==5.10.0; python_version >= '3.7'
iniconfig==2.0.0; python_version >= '3.7'
ipdb==0.13.9; python_version >= '2.7'
ipython==8.22.1; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.3; python_version >= '3.7'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
mako==1.3.2; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.18.0; python_version >= '3.7'
marshmallow-sqlalchemy==0.28.1; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pluggy==1.4.0; python_version >= '3.8'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pygments==2.17.2; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
pytest==7.1.3; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
referencing==0.33.0; python_version >= '3.8'
rpds-py==0.18.0; python_version >= '3.8'
setuptools==69.1.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.27; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
stack-data==0.6.3
toml==0.10.2; python_version >= '3.7'
tomli==2.0.1; python_version >= '3.7'
traitlets==5.14.1; python_version >= '3.8'
typing-extensions==4.10.0; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
